Fix #2
